backend/requirements.txt
Flask==3.0.2
python-dotenv==1.0.1
yagmail==0.15.293
pytest==8.3.2
